+++ libs/ibm/langchain_ibm/toolkit.py
@@ -1,7 +1,6 @@
 """IBM watsonx.ai Toolkit wrapper."""
 
 import urllib.parse
-from copy import deepcopy
 from typing import (
     Any,
     Dict,
@@ -29,7 +28,7 @@ from pydantic import (
 )
 from typing_extensions import Self
 
-from langchain_ibm.utils import check_for_attribute
+from langchain_ibm.utils import check_for_attribute, convert_to_watsonx_tool
 
 
 class WatsonxTool(BaseTool):
@@ -279,78 +278,3 @@ def json_schema_to_pydantic_model(name: str, schema: Dict[str, Any]) -> Type[Bas
         fields[field_name] = (py_type, ... if is_required else None)
 
     return create_model(name, **fields)  # type: ignore[call-overload]
-
-
-def convert_to_watsonx_tool(tool: WatsonxTool) -> dict:
-    """Convert `WatsonxTool` to watsonx tool structure.
-
-    Args:
-        tool: `WatsonxTool` from `WatsonxToolkit`
-
-
-    Example:
-
-    .. code-block:: python
-
-        from langchain_ibm import WatsonxToolkit
-
-        watsonx_toolkit = WatsonxToolkit(
-            url="https://us-south.ml.cloud.ibm.com",
-            apikey="*****",
-        )
-        weather_tool = watsonx_toolkit.get_tool("Weather")
-        convert_to_watsonx_tool(weather_tool)
-
-        # Return
-        # {
-        #     "type": "function",
-        #     "function": {
-        #         "name": "Weather",
-        #         "description": "Find the weather for a city.",
-        #         "parameters": {
-        #             "type": "object",
-        #             "properties": {
-        #                 "location": {
-        #                     "title": "location",
-        #                     "description": "Name of the location",
-        #                     "type": "string",
-        #                 },
-        #                 "country": {
-        #                     "title": "country",
-        #                     "description": "Name of the state or country",
-        #                     "type": "string",
-        #                 },
-        #             },
-        #             "required": ["location"],
-        #         },
-        #     },
-        # }
-
-    """
-
-    def parse_parameters(input_schema: dict | None) -> dict:
-        if input_schema:
-            parameters = deepcopy(input_schema)
-        else:
-            parameters = {
-                "type": "object",
-                "properties": {
-                    "input": {
-                        "description": "Input to be used when running tool.",
-                        "type": "string",
-                    },
-                },
-                "required": ["input"],
-            }
-
-        return parameters
-
-    watsonx_tool = {
-        "type": "function",
-        "function": {
-            "name": tool.name,
-            "description": tool.description,
-            "parameters": parse_parameters(tool.tool_input_schema),
-        },
-    }
-    return watsonx_tool
