+++ libs/ibm/langchain_ibm/utils.py
@@ -1,8 +1,11 @@
 from copy import deepcopy
-from typing import Any, Dict, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional, Union
 
 from ibm_watsonx_ai.foundation_models.schema import BaseSchema  # type: ignore
 from pydantic import SecretStr
+
+if TYPE_CHECKING:
+    from langchain_ibm.toolkit import WatsonxTool
 
 
 def check_for_attribute(value: SecretStr | None, key: str, env_key: str) -> None:
@@ -57,3 +60,78 @@ def check_duplicate_chat_params(params: dict, kwargs: dict) -> None:
             f"Duplicate parameters found in params and keyword arguments: "
             f"{list(duplicate_keys)}"
         )
+
+
+def convert_to_watsonx_tool(tool: "WatsonxTool") -> dict:
+    """Convert `WatsonxTool` to watsonx tool structure.
+
+    Args:
+        tool: `WatsonxTool` from `WatsonxToolkit`
+
+
+    Example:
+
+    .. code-block:: python
+
+        from langchain_ibm import WatsonxToolkit
+
+        watsonx_toolkit = WatsonxToolkit(
+            url="https://us-south.ml.cloud.ibm.com",
+            apikey="*****",
+        )
+        weather_tool = watsonx_toolkit.get_tool("Weather")
+        convert_to_watsonx_tool(weather_tool)
+
+        # Return
+        # {
+        #     "type": "function",
+        #     "function": {
+        #         "name": "Weather",
+        #         "description": "Find the weather for a city.",
+        #         "parameters": {
+        #             "type": "object",
+        #             "properties": {
+        #                 "location": {
+        #                     "title": "location",
+        #                     "description": "Name of the location",
+        #                     "type": "string",
+        #                 },
+        #                 "country": {
+        #                     "title": "country",
+        #                     "description": "Name of the state or country",
+        #                     "type": "string",
+        #                 },
+        #             },
+        #             "required": ["location"],
+        #         },
+        #     },
+        # }
+
+    """
+
+    def parse_parameters(input_schema: dict | None) -> dict:
+        if input_schema:
+            parameters = deepcopy(input_schema)
+        else:
+            parameters = {
+                "type": "object",
+                "properties": {
+                    "input": {
+                        "description": "Input to be used when running tool.",
+                        "type": "string",
+                    },
+                },
+                "required": ["input"],
+            }
+
+        return parameters
+
+    watsonx_tool = {
+        "type": "function",
+        "function": {
+            "name": tool.name,
+            "description": tool.description,
+            "parameters": parse_parameters(tool.tool_input_schema),
+        },
+    }
+    return watsonx_tool
